+++ internal/bandchain/bandtss/bandtss.go
@@ -1,4 +1,4 @@
-package types
+package bandtss
 
 // SigningID is the type-safe unique identifier type for bandtss signing info.
 type SigningID uint64
